优惠啊

+++ src/components/TrashDetail.vue
@@ -53,6 +53,10 @@
       this.getNotebooks()
       this.getTrashNotes().then(() => {
         this.setCurTrashNote({ curTrashNoteId: this.$route.query.noteId })
+        this.$router.replace({
+          path: '/trash',
+          query: { noteId: this.curTrashNote.id }
+        })
       })
     },
 
@@ -70,11 +74,31 @@
       ...mapActions(['checkLogin', 'deleteTrashNote', 'revertTrashNote', 'getTrashNotes', 'getNotebooks']),
 
       onDelete() {
-        this.deleteTrashNote({ noteId: this.curTrashNote.id })
+        this.$confirm('彻底删除后将无法恢复哦~', '彻底删除', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        })
+          .then(() => {
+            return this.deleteTrashNote({ noteId: this.curTrashNote.id })
+          })
+          .then(() => {
+            this.setCurTrashNote()
+            this.$router.replace({
+              path: '/trash',
+              query: { noteId: this.curTrashNote.id }
+            })
+          })
       },
 
       onRevert() {
-        this.revertTrashNote({ noteId: this.curTrashNote.id })
+        this.revertTrashNote({ noteId: this.curTrashNote.id }).then(() => {
+          this.setCurTrashNote()
+          this.$router.replace({
+            path: '/trash',
+            query: { noteId: this.curTrashNote.id }
+          })
+        })
       }
     },
